Create README.md for table of contents
I suggest using the README page for this folder as a table of contents. Unlike your other folders, this one contains three documents. So, you can use the README as a navigational tool for readers moving between each of the three documents. This is a pull request so if your team approves just merge my suggestions.